完成时间显示item

+++ app/src/main/java/com/cleven/clchat/manager/CLMessageManager.java
@@ -63,7 +63,7 @@ public class CLMessageManager {
         /// 消息类型
         message.setMessageType(CLMessageBodyType.MessageBodyType_Text.getTypeName());
 
-        int timeStamp = CLUtils.timeStamp;
+        long timeStamp = CLUtils.timeStamp;
         /// 消息id
         message.setMessageId("" + timeStamp);
         /// 发送状态
